indexer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "indexer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # captioning
    "transformers>=4.38",
    "torch>=2.2",
    "Pillow>=10.3",
    "openai-whisper>=20231117",
    # exif
    "exifread>=3.0",
    "mutagen>=1.47",
    # vectorizer
    "sentence-transformers>=3.0",
    # vector store
    "chromadb>=0.5",
    # CLI
    "click>=8.1",
    # progress
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
indexer = "indexer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/indexer"]
